# [WIP] add apiversion and kind to list subitems