… pkg/api/testing/replication_controller_list_example.json
… staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/generated.pb.go
… staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/zz_generated.deepcopy.go