Fixes https://github.com/kubernetes/kubernetes/issues/3030  (from 2014, would that be a record?)

The current state is ugly, but could be tidied up and tests added.  I put a random test-cmd check in there for proof.

Any concerns about where and how I did this?  It requires 
 1. that scheme recognizes your list item
 2. that the list item is registered in the same GV as your list
 3. that you specify a specific (not generic) list.

I'd probably reconcile the `pkg/api/meta` cycle by pushing these into the `pkg/runtime/scheme` package.

@kubernetes/sig-api-machinery-bugs 
@smarterclayton @liggitt @lavalamp @sttts 

```release-note
nested items in lists (like namespacelist), will have apiversion and kind
```



```json
{
  "kind": "NamespaceList",
  "apiVersion": "v1",
  "metadata": {
    "selfLink": "/api/v1/namespaces",
    "resourceVersion": "726"
  },
  "items": [
    {
      "kind": "Namespace",
      "apiVersion": "v1",
      "metadata": {
        "name": "default",
        "selfLink": "/api/v1/namespaces/default",
        "uid": "a4c47f1b-59dd-11e8-95c8-54e1ad486dd3",
        "resourceVersion": "16",
        "creationTimestamp": "2018-05-17T14:21:52Z"
      },
      "spec": {
        "finalizers": [
          "kubernetes"
        ]
      },
      "status": {
        "phase": "Active"
      }
    },
    {
      "kind": "Namespace",
      "apiVersion": "v1",
      "metadata": {
        "name": "kube-public",
        "selfLink": "/api/v1/namespaces/kube-public",
        "uid": "a4c7ad9f-59dd-11e8-95c8-54e1ad486dd3",
        "resourceVersion": "36",
        "creationTimestamp": "2018-05-17T14:21:52Z"
      },
      "spec": {
        "finalizers": [
          "kubernetes"
        ]
      },
      "status": {
        "phase": "Active"
      }
    },
    {
      "kind": "Namespace",
      "apiVersion": "v1",
      "metadata": {
        "name": "kube-system",
        "selfLink": "/api/v1/namespaces/kube-system",
        "uid": "a4c77664-59dd-11e8-95c8-54e1ad486dd3",
        "resourceVersion": "34",
        "creationTimestamp": "2018-05-17T14:21:52Z"
      },
      "spec": {
        "finalizers": [
          "kubernetes"
        ]
      },
      "status": {
        "phase": "Active"
      }
    }
  ]
}

```

BEFORE
```
[deads@deads-02 kubernetes]$ go test k8s.io/kubernetes/pkg/api/testing -bench=Benchmark
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/api/testing
BenchmarkPodConversion-8                                               	  500000	      2572 ns/op
BenchmarkNodeConversion-8                                              	 1000000	      1761 ns/op
BenchmarkReplicationControllerConversion-8                             	  500000	      3004 ns/op
BenchmarkReplicationControllerListConversion-8                         	  100000	     14848 ns/op
BenchmarkPodCopy-8                                                     	 1000000	      1437 ns/op
BenchmarkNodeCopy-8                                                    	 1000000	      1733 ns/op
BenchmarkReplicationControllerCopy-8                                   	  500000	      2332 ns/op
BenchmarkPodDefaulting-8                                               	 5000000	       342 ns/op
BenchmarkEncodeCodecProtobuf-8                                         	  300000	      3644 ns/op
BenchmarkEncodeCodecFromInternalProtobuf-8                             	  200000	      5279 ns/op
BenchmarkEncodeProtobufGeneratedMarshal-8                              	  500000	      2736 ns/op
BenchmarkDecodeCodecToInternalProtobuf-8                               	  200000	      6998 ns/op
BenchmarkDecodeIntoProtobuf-8                                          	  300000	      4011 ns/op
BenchmarkEncodeCodec-8                                                 	  100000	     17979 ns/op
BenchmarkEncodeCodecFromInternal-8                                     	  100000	     19209 ns/op
BenchmarkEncodeJSONMarshal-8                                           	  100000	     19113 ns/op
BenchmarkDecodeCodec-8                                                 	   30000	     37095 ns/op
BenchmarkDecodeIntoExternalCodec-8                                     	   50000	     35739 ns/op
BenchmarkDecodeIntoInternalCodec-8                                     	   50000	     37947 ns/op
BenchmarkDecodeIntoJSON-8                                              	   30000	     49692 ns/op
BenchmarkDecodeIntoJSONCodecGenConfigFast-8                            	  100000	     13865 ns/op
BenchmarkDecodeIntoJSONCodecGenConfigCompatibleWithStandardLibrary-8   	  100000	     16051 ns/op
BenchmarkToFromUnstructured-8                                          	    3000	    356192 ns/op
BenchmarkToFromUnstructuredViaJSON-8                                   	   10000	    175606 ns/op
PASS

```

AFTER
```
[deads@deads-02 kubernetes]$ go test k8s.io/kubernetes/pkg/api/testing -bench=Benchmark
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/api/testing
BenchmarkPodConversion-8                                               	  500000	      2601 ns/op
BenchmarkNodeConversion-8                                              	 1000000	      1783 ns/op
BenchmarkReplicationControllerConversion-8                             	  500000	      3134 ns/op
BenchmarkReplicationControllerListConversion-8                         	  100000	     16719 ns/op
BenchmarkPodCopy-8                                                     	 1000000	      1450 ns/op
BenchmarkNodeCopy-8                                                    	 1000000	      1664 ns/op
BenchmarkReplicationControllerCopy-8                                   	  500000	      2441 ns/op
BenchmarkPodDefaulting-8                                               	 5000000	       363 ns/op
BenchmarkEncodeCodecProtobuf-8                                         	  300000	      3820 ns/op
BenchmarkEncodeCodecFromInternalProtobuf-8                             	  200000	      5743 ns/op
BenchmarkEncodeProtobufGeneratedMarshal-8                              	  500000	      2878 ns/op
BenchmarkDecodeCodecToInternalProtobuf-8                               	  200000	      7500 ns/op
BenchmarkDecodeIntoProtobuf-8                                          	  300000	      4217 ns/op
BenchmarkEncodeCodec-8                                                 	  100000	     18871 ns/op
BenchmarkEncodeCodecFromInternal-8                                     	  100000	     19575 ns/op
BenchmarkEncodeJSONMarshal-8                                           	  100000	     19359 ns/op
BenchmarkDecodeCodec-8                                                 	   30000	     36683 ns/op
BenchmarkDecodeIntoExternalCodec-8                                     	   50000	     35289 ns/op
BenchmarkDecodeIntoInternalCodec-8                                     	   50000	     37216 ns/op
BenchmarkDecodeIntoJSON-8                                              	   30000	     50176 ns/op
BenchmarkDecodeIntoJSONCodecGenConfigFast-8                            	  100000	     14181 ns/op
BenchmarkDecodeIntoJSONCodecGenConfigCompatibleWithStandardLibrary-8   	  100000	     15932 ns/op
BenchmarkToFromUnstructured-8                                          	    3000	    359906 ns/op
BenchmarkToFromUnstructuredViaJSON-8                                   	   10000	    182426 ns/op
PASS
ok  	k8s.io/kubernetes/pkg/api/testing	45.824s

```