@lavalamp 2018-05-17
How badly does this hurt performance? LGTM