@deads2k 2018-05-17
I'll find the benchmark and run it.  How much are we willing to tolerate to make life easier for every client?  Some percentage of the entire encoding path I'd imagine