@lavalamp 2018-05-17
I think I'm willing to pay 5%, but I'm not sure about the scalability folks.

On Thu, May 17, 2018 at 2:07 PM David Eads <notifications@github.com> wrote:

> I'll find the benchmark and run it. How much are we willing to tolerate to
> make life easier for every client? Some percentage of the entire encoding
> path I'd imagine
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/63972#issuecomment-390011582>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAngloEQVXIeDM5oIWC2OdAUUfUya0SRks5tzeaMgaJpZM4UDNGT>
> .
>