@deads2k 2018-05-17
> I had the same question, it looks like a really complex inner loop. I'm imagining this running on every item listed across all resources on informer lists.


Remember that this is only down the conversion path.  Anything on external types doesn't get converted, just decoded.  This won't impact the controllers on the read side.  The real hit is server-side, where we already walk the list multiple times.



> I think I'm willing to pay 5%, but I'm not sure about the scalability folks.

5% seems fair.  It's a big behavior benefit.

@wojtek-t @shyamjvs is there a good existing benchmark in the code I should be looking at for a conversion change.  We'll hit this on the api-server when serving a list.