@shyamjvs 2018-05-18
> is there a good existing benchmark in the code I should be looking at for a conversion change.

@deads2k I'm not sure what you're exactly looking for, but there a whole bunch of benchmarks under `pkg/api/testing` (especially `conversion_test.go`) that you might find useful.

> I think I'm willing to pay 5%, but I'm not sure about the scalability folks.

IIUC that you mean 5% increase in list latencies, that might be ok. We have some slack b/w our current latencies and SLO for list calls (which is atm 5s for namespaced list calls and 10s for cluster-scoped ones).
cc @wojtek-t