@deads2k 2018-05-21
No real difference for normal types, but lists are a more significant bump:

```
OLD BenchmarkReplicationControllerListConversion-8                         	  100000	     14848 ns/op
NEW BenchmarkReplicationControllerListConversion-8                         	  100000	     16719 ns/op
```