@deads2k 2018-05-21
Alright, got a POC commit that shows how to do this fast enough.


```
with specific methods BenchmarkReplicationControllerListConversion-8   	  100000	     14819 ns/op
```