@davidxia 2018-11-05
Hey, are there any updates on this? I'd love to see this fixed as it's convenient to have the Kind and APIVersion for each resource. Let me know how I can help move this along.